+++ src/main/ruby/truffleruby/core/posix.rb
@@ -221,7 +221,8 @@ module Truffle::POSIX
   attach_function :truffleposix_rewinddir, [:pointer], :void, LIBTRUFFLEPOSIX
   attach_function :rmdir, [:string], :int
   attach_function :seekdir, [:pointer, :long], :void
-  attach_function :truffleposix_select, [:int, :pointer, :int, :pointer, :int, :pointer, :long], :int, LIBTRUFFLEPOSIX
+  select_args = [:int, :pointer, :int, :pointer, :int, :pointer, :long]
+  attach_function :truffleposix_select, select_args, :int, LIBTRUFFLEPOSIX
   attach_function :truffleposix_stat, [:string, :pointer], :int, LIBTRUFFLEPOSIX
   attach_function :truffleposix_stat_mode, [:string], :mode_t, LIBTRUFFLEPOSIX
   attach_function :truffleposix_stat_size, [:string], :long, LIBTRUFFLEPOSIX
@@ -232,6 +233,14 @@ module Truffle::POSIX
   attach_function :unlink, [:string], :int
   attach_function :truffleposix_utimes, [:string, :long, :int, :long, :int], :int, LIBTRUFFLEPOSIX
   attach_function :write, [:int, :pointer, :size_t], :ssize_t, LIBC, true
+
+  Truffle::Boot.delay do
+    if NATIVE
+      # We should capture the non-lazy method
+      attach_function_eagerly :truffleposix_select, select_args, :int, LIBTRUFFLEPOSIX, false, :truffleposix_select, self
+      SELECT = method(:truffleposix_select)
+    end
+  end
 
   # Process-related
   attach_function :getegid, [], :gid_t
@@ -310,8 +319,6 @@ module Truffle::POSIX
   unless Truffle::Platform.darwin?
     attach_function :dup3, [:int, :int, :int], :int
   end
-
-  SELECT = method(:truffleposix_select)
 
   def self.with_array_of_ints(ints)
     if ints.empty?